add homework 3

+++ src/main/java/co/istad/chhaya/itespringrestapi/domain/Coffee.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -14,8 +16,7 @@ public class Coffee {
     private Integer id;
     private String name;
     private String description;
-    private Double price;
-
+    private BigDecimal price;
 
 
 }
